Improve decoder/generator training parity.
Previously, the decoder trained on very different examples then what
it received from the generator in the combined model.

It previouslly received the entire model generated seed sequence +
an additional 40 timesteps of extrapoloated data.

This has been corrected such that the decoder only trains on the
first 40 timesteps (the re-generated seed sequence + 1 timestep).